Make package compatible with web-tools 9.0
