+++ README.md
@@ -1,5 +1,5 @@
-# web-tools-builder-ngtemplates
+# web-tools-pipeline-ngtemplates
 
-**web-tools-builder-ngtemplates** is a builder pipeline intended for projects using the [Web-tools](https://github.com/imgix/web-tools) build system. This builder lints and concatenates standalone Angular template files into a single JS file that injects the templates directly into Angular's template cache.
+**web-tools-pipeline-ngtemplates** is a builder pipeline intended for projects using the [Web-tools](https://github.com/imgix/web-tools) build system. This pipeline lints and concatenates standalone Angular template files into a single JS file that injects the templates directly into Angular's template cache.
