+++ tools/web-tools-pipeline-ngtemplates.js
@@ -1,0 +1,3 @@
+module.exports = function setUpNgTemplatesPipeline(gulp) {
+  gulp.pipelineCache.put('ngtemplates', require('./pipelines/pipeline.ngtemplates.js'));
+};
